knowledge-base/requirements.txt
langchain
sentence-transformers
pymilvus
neo4j
llama-cpp-python
pypdf2
python-docx
beautifulsoup4
pytesseract
pillow
pandas
numpy